requirements-vercel.txt
Django==5.2.3
django-tinymce==3.7.1
Pillow==11.3.0
whitenoise==6.9.0
python-decouple==3.8
django-cors-headers==4.7.0
django-storages==1.14.2
psycopg2==2.9.10
dj-database-url==2.1.0
